+++ app/jobs/delete_stale_journeys_job.rb
@@ -1,0 +1,8 @@
+class DeleteStaleJourneysJob < ApplicationJob
+  queue_as :default
+  sidekiq_options retry: 1
+
+  def perform
+    DeleteStaleJourneys.new.call
+  end
+end