+++ app/services/save_answer.rb
@@ -19,6 +19,7 @@ class SaveAnswer
 
     if answer.valid?
       answer.save
+      answer.step.journey.freshen!
       ToggleAdditionalSteps.new(step: answer.step).call
       result.success = true
     end