+++ db/schema.rb
@@ -43,6 +43,10 @@ ActiveRecord::Schema.define(version: 2021_04_01_100528) do
     t.jsonb "section_groups"
     t.uuid "user_id"
     t.index ["user_id"], name: "index_journeys_on_user_id"
+    t.boolean "started", default: true
+    t.index ["started"], name: "index_journeys_on_started"
+    t.datetime "last_worked_on"
+    t.index ["last_worked_on"], name: "index_journeys_on_last_worked_on"
   end
 
   create_table "long_text_answers", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|