+++ spec/factories/journey.rb
@@ -3,6 +3,8 @@ FactoryBot.define do
     category { "catering" }
     liquid_template { "Your answer was {{ answer_47EI2X2T5EDTpJX9WjRR9p }}" }
     section_groups { [] }
+    started { true }
+    last_worked_on { Time.zone.now }
 
     association :user, factory: :user