+++ spec/models/journey_spec.rb
@@ -65,4 +65,29 @@ RSpec.describe Journey, type: :model do
       end
     end
   end
+
+  describe "freshen!" do
+    it "set started to true" do
+      journey = build(:journey, :catering)
+      journey.freshen!
+      expect(journey.reload.started).to eq(true)
+    end
+
+    it "sets the last_worked_on to now" do
+      travel_to Time.zone.local(2004, 11, 24, 1, 4, 44)
+      journey = build(:journey, :catering)
+
+      journey.freshen!
+
+      expect(journey.last_worked_on).to eq(Time.zone.now)
+    end
+
+    context "when started is already true" do
+      it "does not update the record" do
+        journey = build(:journey, :catering, started: true)
+        expect(journey).not_to receive(:update).with(started: true)
+        journey.freshen!
+      end
+    end
+  end
 end